Use a venv to manage dependencies
Update puzzle.Parser,parse_points to explicitly specify the type of the integer array, otherwise we get 64b integers instead of the desired 32b

+++ lib/puzzler/puzzle.py
@@ -73,7 +73,7 @@ class Parser:
         return Puzzle.Piece.Source(data['id'], tuple(data['rect']))
 
     def parse_points(self, data):
-        pts = np.array(puzzler.chain.ChainCode().decode(data))
+        pts = np.array(puzzler.chain.ChainCode().decode(data), dtype=np.int32)
         pts.setflags(write=0)
         return pts
